test: migrate mocha test
