+++ test/cases/mocha/main.mjs
@@ -1,0 +1,2 @@
+export const main = () => "main";
+export const mainAsync = async () => await "main";
