+++ test/cases/mocha/mocha/suite-1.subset.yaml
@@ -1,0 +1,1 @@
+{}
