+++ test/cases/mocha/mocha/suite-2.subset.yaml
@@ -1,0 +1,1 @@
+{}
