+++ test/cases/mocha/mocha/suite.subset.yaml
@@ -1,0 +1,1 @@
+{}
